agregado microservicios de main

+++ ms-clase/src/main/java/com/example/ms_clase/service/ClaseService.java
@@ -7,6 +7,7 @@ import com.example.ms_clase.dto.ClaseResponse;
 import com.example.ms_clase.model.Clase;
 import com.example.ms_clase.repository.ClaseRepository;
 import jakarta.persistence.EntityNotFoundException;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -27,6 +28,11 @@ public class ClaseService {
     public ClaseResponse add(ClaseRequest c,String token){
         log.info("Anadir Clase", keyValue("Nombre de Clase", c.getNombreClase()));
         var entrenador = entrenadorClient.getEntrenador(c.getIdEntrenador(), token);
+        if(entrenador == null){
+            log.warn("Entrenador no encontrado",
+                    keyValue("idEntrenador", c.getIdEntrenador()));
+            throw new EntityNotFoundException("Entrenador no encontrado");
+        }
         Clase clase1 = new Clase();
         clase1.setCupos(c.getCupos());
         clase1.setNivelDeClase(c.getNivelDeClase());
@@ -36,8 +42,12 @@ public class ClaseService {
         clase1.setFechaRealizacion(c.getFechaRealizacion());
         clase1.setHoraRealizacion(c.getHoraRealizacion());
         clase1.setIdEntrenador(c.getIdEntrenador());
+        Clase saveClase =repository.save(clase1);
+        log.info("Clase creada correctamente",
+                keyValue("idClase", saveClase.getId())
+        );
 
-        return mapToResponse(repository.save(clase1),token);
+        return mapToResponse(saveClase,token);
     }
 
     public ClaseResponse findById(Long id,String token){
@@ -50,15 +60,26 @@ public class ClaseService {
     }
 
     public List<ClaseResponse> getAll(String token){
-
+        log.info("Listando clases");
         return repository.findAll().stream()
                 .map(c -> mapToResponse(c,token))
                 .toList();
     }
 
     public ClaseResponse update(Long id,ClaseRequest c,String token){
+        log.info("Actualizando clase",
+                keyValue("idClase", id)
+        );
         Clase clase1 = repository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Clase no encontrada"));
+        var entrenador = entrenadorClient.getEntrenador(c.getIdEntrenador(), token);
+
+        if(entrenador == null){
+            log.warn("Entrenador no encontrado",
+                    keyValue("idEntrenador", c.getIdEntrenador())
+            );
+            throw new EntityNotFoundException("Entrenador no encontrado");
+        }
         clase1.setCupos(c.getCupos());
         clase1.setNivelDeClase(c.getNivelDeClase());
         clase1.setNombreClase(c.getNombreClase());
@@ -67,15 +88,33 @@ public class ClaseService {
         clase1.setFechaRealizacion(c.getFechaRealizacion());
         clase1.setHoraRealizacion(c.getHoraRealizacion());
         clase1.setIdEntrenador(c.getIdEntrenador());
-        return mapToResponse(repository.save(clase1),token);
+        Clase updateClase = repository.save(clase1);
+        log.info("Clase actualizada correctamente",
+                keyValue("idClase", updateClase.getId())
+        );
+        return mapToResponse(updateClase,token);
     }
 
     public void delete(Long id ){
-        repository.deleteById(id);
+        log.info("Eliminando clase",
+                keyValue("idClase", id)
+        );
 
+        if (!repository.existsById(id)) {
+            log.warn("clase a eliminar inexistente",
+                    keyValue("idClase", id)
+            );
+            throw new EntityNotFoundException("No se puede eliminar clase no encontrada");
+        }
+        repository.deleteById(id);
+        log.info("Clase eliminada correctamente",
+                keyValue("idClase", id)
+        );
     }
     private ClaseResponse mapToResponse(Clase c, String token) {
-
+        log.info("Mapeando clase",
+                keyValue("idClase", c.getId())
+        );
         var entrenador1 = entrenadorClient.getEntrenador(c.getIdEntrenador(), token);
 
         return ClaseResponse.builder().id(c.getId()).nombreClase(c.getNombreClase()).
@@ -89,4 +128,71 @@ public class ClaseService {
                 build();
 
     }
+
+    @Transactional
+    public void personaInscrita(Long id){
+        /*Clase clase = repository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Clase no encontrada"));;
+        if (clase.getCupos() <= 0) {
+            throw new RuntimeException("No quedan cupos disponibles para esta clase");
+        }
+
+        clase.setCupos(clase.getCupos() - 1 );
+        log.info("Cupo actualizado", keyValue("idClase", clase.getId()));
+        repository.save(clase);*/
+
+        log.info("resta de cupo ", keyValue("idClase", id));
+
+        int filasAfectadas = repository.restarCupo(id);
+
+        if (filasAfectadas == 0) {
+            log.warn("No se pudo restar cupo",
+                    keyValue("idClase", id)
+            );
+
+            throw new RuntimeException("No se pudo restar el cupo");
+        }
+
+        log.info("Cupo actualizado correctamente",
+                keyValue("idClase", id)
+        );
+
+    }
+    @Transactional
+    public void removerInscripcion(Long id){
+        /*Clase clase = repository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Clase no encontrada"));;
+
+        clase.setCupos(clase.getCupos() + 1 );
+        log.info("Cupo actualizado", keyValue("idClase", clase.getId()));
+        repository.save(clase);*/
+        log.info("Sumando cupo de clase",
+                keyValue("idClase", id)
+        );
+
+        int filasAfectadas = repository.sumarCupo(id);
+
+        if (filasAfectadas == 0) {
+            log.warn("No se pudo sumar cupo",
+                    keyValue("idClase", id)
+            );
+
+            throw new RuntimeException("No se pudo sumar el cupo");
+        }
+
+        log.info("Cupo restaurado correctamente",
+                keyValue("idClase", id)
+        );
+    }
+
+
+    public List<ClaseResponse> buscarPorNombre(String nombre,String token){
+        log.info("buscando clases por nombre ",keyValue("nombre",nombre));
+        List<Clase> clases = repository.findByNombreClase(nombre);
+        log.info("Se devuelve lista de clases por nombre",keyValue("nombre",nombre));
+        return clases.stream()
+                .map(c -> mapToResponse(c,token))
+                .toList();
+    }
+
 }
